Add option to error/timeout if the pool is full
Currently, if you attempt to acquire a resource and the pool is full, the
callback will not run until a space becomes available. This can lead to a
deadlock if a pool member's release is blocked on the acquire callback running.

To mitigate the deadlock possibility, support a third argument
`pool.acquire(cb, priority, {timeout: 5000})`, which will hit the callback with
an error if the pool is still full after 5000ms. Adds some tests that this
behaves as expected.

Adds a helper function pool._isFull, which checks both the _count and the list
of _availableObjects - I think we have to check both to ensure that the pool is
full, and that this abstraction doesn't exist yet.

Callers may want to wait for a small amount of time before giving up on the
pool; I'm going to add that option in a second commit.

The `options` dictionary for pool.acquire() now takes a second argument,
`timeout`. Pass an integer number of milliseconds to cancel the acquire() call
and hit the callback with pool.Full after the given amount of time has elapsed.
Hits the callback with an immediate error if the timeout is a negative number,
zero, or NaN.

Adds a number of tests that this logic behaves as expected.

This work was sponsored by [Shyp](https://shyp.com).